+++ src/macros.rs
@@ -1,4 +1,4 @@
-use crate::converter::{convert_type_to_wgpu, has_type};
+use crate::converter::{convert_type_to_wgpu, has_type, convert_mat_type_to_wgou};
 use crate::parser::TokenVertexFormat;
 use crate::parser::parse_attrs;
 
@@ -10,7 +10,8 @@ struct Entity {
 #[derive(Debug)]
 struct EntityFieldsAttrs {
     name: String,
-    data: u32
+    data: u32,
+    ty: Option<String>
 }
 
 #[derive(Debug)]
@@ -24,11 +25,25 @@ fn get_entity_field(field: &syn::Field) -> Option<EntityFields> {
     let mut attrs: Vec<EntityFieldsAttrs> = Vec::new();
 
     parse_attrs(&field.attrs,Box::new(|attr| {
-        let lint : syn::LitInt = attr.attribute.parse_args().expect("Only integer is authorize for shader location data");
+        let name = attr.segment.ident.to_string();
+        if name.starts_with("mat") {
+            let mat : crate::parser::AttrMat = attr.attribute.parse_args().unwrap();
+            
+            attrs.push(EntityFieldsAttrs {
+                name,
+                data: mat.data,
+                ty: Some(mat.ident.to_string()),
+            });
 
+            return
+        }
+
+        let lint : syn::LitInt = attr.attribute.parse_args().expect("Only integer is authorize for shader location data");
+    
         attrs.push(EntityFieldsAttrs {
             name: attr.segment.ident.to_string(),
-            data: lint.base10_parse().unwrap()
+            data: lint.base10_parse().unwrap(),
+            ty: None
         });
     }));
 
@@ -39,6 +54,30 @@ fn get_entity_field(field: &syn::Field) -> Option<EntityFields> {
     };
 
     Some(entity_fields)
+}
+
+fn process_wgpu_type(
+    format: &crate::converter::WGPUData, 
+    shader_locations: &mut Vec<u32>,
+    attrs: &mut Vec<proc_macro2::TokenStream>,
+    offset: &u64
+) {
+    let tty = TokenVertexFormat { attribute: format.wgpu_type.ty};
+    let shader_location = format.shader_location;
+
+    if shader_locations.contains(&shader_location) {
+        panic!("Cannot have two time the same location in the same struct");
+    }
+
+    shader_locations.push(shader_location);
+
+    attrs.push(quote::quote! {
+        wgpu::VertexAttribute {
+            offset: #offset,
+            format: #tty,
+            shader_location: #shader_location
+        }
+    });
 }
 
 fn require_repr_c(attrs : &std::vec::Vec<syn::Attribute>) {
@@ -56,7 +95,7 @@ fn require_repr_c(attrs : &std::vec::Vec<syn::Attribute>) {
     }
 }
 
-pub fn derive_wrld_desc(item: proc_macro::TokenStream) -> proc_macro::TokenStream {
+pub fn derive_wrld_desc(item: proc_macro::TokenStream, step_mode: wgpu::VertexStepMode) -> proc_macro::TokenStream {
     let syn::DeriveInput {ident, data, attrs, ..} = syn::parse_macro_input!(item as syn::DeriveInput);
     let fields = if let syn::Data::Struct(syn::DataStruct {
         fields: syn::Fields::Named(syn::FieldsNamed { ref named, ..}),
@@ -81,34 +120,33 @@ pub fn derive_wrld_desc(item: proc_macro::TokenStream) -> proc_macro::TokenStrea
 
     for i in entity.fields {
         for attr in i.attrs {
-            let format = convert_type_to_wgpu(attr.name.as_str(), attr.data);
-            let tty = TokenVertexFormat { attribute: format.wgpu_type.ty};
-            let shader_location = format.shader_location;
+            if attr.ty == None {
+                let format = convert_type_to_wgpu(&attr.name, attr.data).unwrap();
+                process_wgpu_type(&format, &mut shader_locations, &mut attrs, &offset);
+                offset += format.wgpu_type.offset;
+            } else {
+                let mat_format = convert_mat_type_to_wgou(
+                    &attr.name, 
+                    attr.data,
+                    &mut attr.ty.unwrap()
+                );
 
-            if shader_locations.contains(&shader_location) {
-                panic!("Cannot have two time the same location in the same struct");
-            }
-
-            shader_locations.push(shader_location);
-
-            attrs.push(quote::quote! {
-                wgpu::VertexAttribute {
-                    offset: #offset,
-                    format: #tty,
-                    shader_location: #shader_location
+                for format in mat_format {
+                    process_wgpu_type(&format, &mut shader_locations, &mut attrs, &offset);
+                    offset += format.wgpu_type.offset;
                 }
-            });
-
-            offset += format.wgpu_type.offset;
+            }
         }
     }
+
+    let step_mode = crate::parser::TokenVertexStepMode {step_mode};
 
     quote::quote! {
         impl #ident {
             pub fn desc<'a>() -> wgpu::VertexBufferLayout<'a> {
                 wgpu::VertexBufferLayout {
                     array_stride: #offset as wgpu::BufferAddress,
-                    step_mode: wgpu::VertexStepMode::Vertex,
+                    step_mode: #step_mode,
                     attributes: &[#(#attrs),*]
                 }
             }
@@ -116,7 +154,7 @@ pub fn derive_wrld_desc(item: proc_macro::TokenStream) -> proc_macro::TokenStrea
     }.into()
 }
 
-pub fn derive_wrsl_buffer_data(item: proc_macro::TokenStream) -> proc_macro::TokenStream {
+pub fn derive_wrld_buffer_data(item: proc_macro::TokenStream) -> proc_macro::TokenStream {
     let syn::DeriveInput {ident, data, ..} = syn::parse_macro_input!(item as syn::DeriveInput);
     let fields = if let syn::Data::Struct(syn::DataStruct {
         fields: syn::Fields::Named(syn::FieldsNamed { ref named, ..}),